Split tool definitions out into separate files
